added /load get request to launch execution. split loading code from scheduler into Loader component. springboot -> 2.4.5
